Docs: Correct `@return` type for `WP_Block_Parser::parse()`.
This affects:
* `WP_Block_Parser::parse()`
* `WP_Block_Parser::$output`

Both the method and the property are documented as returning `WP_Block_Parser_Block[]` (an array of `WP_Block_Parser_Block` objects), but the result is in fact an array of arrays of various values, so `array[]` is the correct notation.

Follow-up to [43751] for the 5.0 branch, [44116] for trunk.

Props Chouby.
Fixes #56581.

git-svn-id: https://develop.svn.wordpress.org/trunk@54194 602fd350-edb4-49c9-b593-d223f7449a82