Customize: Fix a wrong condition on `WP_Customize_Manager::has_published_pages()`.
The Homepage Settings section on the customizer uses `WP_Customize_Manager::has_published_pages()` as `active_callback`. [53057] introduced an error on the logical check when the queried page is not a top-level page, which leads `has_published_pages` to incorrectly return an empty array.

This changeset fixes the issue by adding `'hierarchical' => 0` to `get_pages()` arguments.

Follow-up to [53057].

Props felipelavinz, dlh, costdev.
Fixes #57198.



git-svn-id: https://develop.svn.wordpress.org/trunk@54935 602fd350-edb4-49c9-b593-d223f7449a82