Pings/Trackbacks: Remove a mention of the "Page" post type from the Discussion meta box.
This changeset removes a mention of the "Page" post type from the Discussion meta box as this is a generic meta box, used for multiple post types. It makes it more consistent with the other option available in the meta box ("Allow comments") and also with the corresponding option in the block editor.

Follow-up to [12323].

Props jeremyfelt, sabernhardt, audrasjb, virgar.
Fixes #57429.
See #11346.


git-svn-id: https://develop.svn.wordpress.org/trunk@55040 602fd350-edb4-49c9-b593-d223f7449a82